dbstealerworm_stateHandlerA0A 98.33->98.94: s16 return type for Obj_GetYawDeltaToObject drops extsh on rotX += RMW
(cherry picked from commit 4db42ef0ba0f4916b20dc4c236976ccfd1cb377d)

+++ src/main/dll/dll_0242_dbstealerworm.c
@@ -1434,7 +1434,7 @@ int dbstealerworm_stateHandlerA0A(int obj, int p2)
     extern int Stack_IsFull(int sp);
     extern void Stack_Push(int sp, int* args);
     extern void ObjMsg_SendToObject(int, int, int, int);
-    extern int Obj_GetYawDeltaToObject(int, int, f32*);
+    extern s16 Obj_GetYawDeltaToObject(int, int, f32*);
     extern f32 lbl_803E6310;
     extern f32 lbl_803E6314;
     extern f32 lbl_803E6318;
